feat: refactoring create_workload_template to kubernetes_manager module
