+++ app/v1/modules/kubernetes_manager/__init__.py
@@ -3,9 +3,13 @@
 
 from .script_create_environment_values_file import ScriptEnvironValuesFileCreator
 from .script_create_vars_file import ScriptVarsFileCreator
+from .script_create_workload_template import ScriptWorkloadTemplateCreator
+from .script_create_workload import ScriptWorkloadCreator
 
 
 __all__ = [
     "ScriptEnvironValuesFileCreator",
     "ScriptVarsFileCreator",
+    "ScriptWorkloadTemplateCreator",
+    "ScriptWorkloadCreator",
 ]
