+++ app/v2/main/helm_chart_helper_v1.py
@@ -8,17 +8,10 @@ from kubernetes import client
 from kubernetes import config
 
 # Refactor Imports
+from app.v1.modules.kubernetes_manager import ScriptWorkloadCreator
 from app.v1.modules.kubernetes_services import ScriptAnnotationsReaderService
-from app.v1.modules.kubernetes_services import ScriptEnvironReaderService
-from app.v1.modules.kubernetes_services import ScriptEnvironReaderFromService
-from app.v1.modules.kubernetes_services import ScriptHostAliasesReaderService
-from app.v1.modules.kubernetes_services import ScriptImagePullSecretsReaderService
 from app.v1.modules.kubernetes_services import ScriptIngressRulesReaderService
 from app.v1.modules.kubernetes_services import ScriptIngressTlsReaderService
-from app.v1.modules.kubernetes_services import ScriptServicesCreatorService
-from app.v1.modules.kubernetes_services import ScriptToDictParserService
-from app.v1.modules.kubernetes_services import ScriptVolumesReaderService
-from app.v1.modules.kubernetes_services import ScriptVolumeMountsReaderService
 
 # Global variable
 app_version = "version1"
@@ -166,7 +159,7 @@ def load_kubernetes_data(conf):
                 pass
         elif kind in ['Job', 'Deployment', 'StatefulSet']:
             for component in conf['components'][kind]:
-                values[component] = create_workload(
+                values[component] = ScriptWorkloadCreator.create_workload(
                     kind=kind,
                     name=component,
                     k8s_client=client,
@@ -188,40 +181,6 @@ def load_kubernetes_data(conf):
                 pass
         conf['kubernetes']['values'][kind] = values
     pass
-
-
-def create_workload_template(ret, name):
-    return dict(
-        annotations=ScriptAnnotationsReaderService.read_annotations(ret.items[0].metadata.annotations),
-        selectorLabels=dict(
-            app=name
-        ),
-        image=dict(
-            repository=ret.items[0].spec.template.spec.containers[0].image.split(':')[0],
-            tag=ret.items[0].spec.template.spec.containers[0].image.split(':')[1]
-        ),
-        command=ret.items[0].spec.template.spec.containers[0].command,
-        args=ret.items[0].spec.template.spec.containers[0].args,
-        env=ScriptEnvironReaderService.read_env(ret.items[0].spec.template.spec.containers[0].env),
-        envFrom=ScriptEnvironReaderFromService.read_env_from(ret.items[0].spec.template.spec.containers[0].env_from),
-        service=dict(
-            ports=ScriptServicesCreatorService.create_services(ret.items[0].spec.template.spec.containers[0].ports)
-        ),
-        volumes=ScriptVolumesReaderService.read_volumes(ret.items[0].spec.template.spec.volumes),
-        volumeMounts=ScriptVolumeMountsReaderService.read_volume_mounts(ret.items[0].spec.template.spec.containers[0].volume_mounts),
-        serviceAccount=(ret.items[0].spec.template.spec.service_account, None)[
-            ret.items[0].spec.template.spec.service_account == 'default'
-        ],
-        imagePullSecrets=ScriptImagePullSecretsReaderService.read_image_pull_secrets(ret.items[0].spec.template.spec.image_pull_secrets),
-        hostAliases=ScriptHostAliasesReaderService.read_host_aliases(ret.items[0].spec.template.spec.host_aliases),
-        readinessProbe=ScriptToDictParserService.to_dict(ret.items[0].spec.template.spec.containers[0].readiness_probe),
-        livenessProbe=ScriptToDictParserService.to_dict(ret.items[0].spec.template.spec.containers[0].liveness_probe),
-        # TODO: Add missing resources
-        # securityContext=to_dict(ret.items[0].spec.template.spec.containers[0].security_context),
-        # strategy
-        # resources
-        # security_context
-    )
 
 
 def create_configmap_or_secret(kind, name, k8s_client, namespace):
@@ -246,32 +205,6 @@ def create_configmap_or_secret(kind, name, k8s_client, namespace):
             string_data[item] = base64.b64decode(ret.items[0].data[item]).decode("utf-8")
         result['stringData'] = string_data
     return result
-
-
-def create_workload(kind, name, k8s_client, namespace):
-    v1 = k8s_client.AppsV1Api()
-    result = dict()
-    print(name)
-    ret = ''
-    if kind == "Job":
-        v1 = k8s_client.BatchV1Api()
-        ret = v1.list_namespaced_job(
-            field_selector="metadata.name={name}".format(name=name),
-            namespace=namespace
-        )
-    elif kind == "StatefulSet":
-        ret = v1.list_namespaced_stateful_set(
-            field_selector="metadata.name={name}".format(name=name),
-            namespace=namespace
-        )
-        result = result | dict(replicas=ret.items[0].spec.replicas)
-    elif kind == "Deployment":
-        ret = v1.list_namespaced_deployment(
-            field_selector="metadata.name={name}".format(name=name),
-            namespace=namespace
-        )
-        result = result | dict(replicas=ret.items[0].spec.replicas)
-    return result | create_workload_template(ret, name)
 
 
 def create_ingress(name: str, k8s_client, namespace, name_suffix=''):
